// packages/ozon/src/endpoints/notification/shared/v1/types.ts
export type OzonNotificationType =
  | "TYPE_NEW_MESSAGE"
  | "TYPE_UPDATE_MESSAGE"
  | "TYPE_MESSAGE_READ"
  | "TYPE_CHAT_CLOSED"
  | "TYPE_NEW_POSTING"
  | "TYPE_POSTING_CANCELLED"
  | "TYPE_STATE_CHANGED"
  | "TYPE_DELIVERY_DATE_CHANGED"
  | "TYPE_CUTOFF_DATE_CHANGED"
  | "TYPE_CREATE_ITEM"
  | "TYPE_UPDATE_ITEM"
  | "TYPE_CREATE_OR_UPDATE_ITEM"
  | "TYPE_STOCKS_CHANGED"
  | "TYPE_FBO_POSTING_NEW"
  | "TYPE_FBO_POSTING_CANCELLED"
  | "TYPE_FBO_POSTING_STATE_CHANGED"
  | "TYPE_FBO_POSTING_DELIVERY_DATE_CHANGED"
  | "TYPE_FBO_STOCKS_CHANGED"
  | "TYPE_ORDER_NEW"
  | "TYPE_ORDER_CANCELLED"
  | "TYPE_ORDER_STATE_CHANGED";

export type OzonNotificationCheckErrorType =
  | "REQUEST_ERROR"
  | "REQUEST_TIMEOUT"
  | "SERVER_FAULT"
  | "STATUS_CODE_NOT_OK"
  | "EMPTY_BODY"
  | "INVALID_BODY"
  | "INVALID_JSON"
  | "WRONG_RESULT_FIELD"
  | "WRONG_RESULT_TIME_FIELD";